refactor: split RecommendationPagingSources
